requirements.txt
langchain
langchain-community
langchain-google-genai
streamlit
ipykernel
python-dotenv
ipywidgets
PyPDF2

-e .